added consumer example
